Добавляет возможность отрисовки по 5 комментов
