+++ js/data.js
@@ -1,0 +1,50 @@
+// Описание фотографии.
+const DESCRIPTIONS = [
+  'Аккуратный пляж',
+  'Путь на пляж',
+  'Вид на океан',
+  'Фотографиня',
+  'Релакс',
+  'Новое приобретение',
+  'Мой завтрак',
+  'Регидратация',
+  'Привет лунатикам!',
+  'Обувница',
+  'Путь в никуда',
+  'Игра на контрастах',
+  'А ужин отдай врагу',
+  'Гуанчжоу',
+  'Когда устали лапки',
+  'Вид из иллюминатора',
+  'Культурно просвещаемся',
+  'Подарок от деда',
+  'Мечта шахтёра',
+  'Городской пейзаж в жёлтых тонах',
+  'Том Ям на минималках',
+  'Закат',
+  'Не насекомое явно...',
+  'Show must go on',
+  'Сафари-парк',
+];
+
+// Примеры комментариев.
+const MESSAGE_LINES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+// Имена авторов
+const NAMES = [
+  'Ксюся',
+  'Марина Александровна',
+  'Люда',
+  'Ритта',
+  'Серёжа',
+  'Веля',
+];
+
+export { DESCRIPTIONS, MESSAGE_LINES, NAMES };
